ML Zoomcamp 2025/Module9/pyproject.toml
[project]
name = "module9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
]
